replaced 3D ResNet by small manually defined CNN for 2D dataset
