+++ application/test_jobs/minimal_training/app/config/config_fed_client.conf
@@ -75,11 +75,11 @@
     path = "nvflare.app_opt.pt.file_model_persistor.PTFileModelPersistor"
     args {
       model {
-        path = "models.resnet.ResNet"
-                      args {
-      in_ch = 1
-      out_ch = 1
-    }
+        path = "models.mini_model.MiniCNNForTesting"
+        args {
+          in_ch = 1
+          out_ch = 1
+        }
       }
     }
   }
