+++ application/test_jobs/minimal_training/app/custom/data/datamodules/__init__.py
@@ -1,7 +1,3 @@
-"""
-This package initializes the necessary modules and classes for the project.
-"""
-
 from .datamodule import DataModule
 
 __all__ = ['DataModule']
