+++ application/test_jobs/minimal_training/app/custom/data/datasets/__init__.py
@@ -1,7 +1,3 @@
-"""
-This package initializes the necessary modules and classes for the project.
-"""
-
 from .mini_dataset_for_testing import MiniDatasetForTesting
 
 __all__ = [MiniDatasetForTesting]
