+++ application/test_jobs/minimal_training/app/custom/models/__init__.py
@@ -1,12 +1,4 @@
-"""
-This package initializes the necessary modules and classes for the project.
-
-Modules:
-    base_model: Contains basic models including VeryBasicModel, BasicModel, and BasicClassifier.
-    resnet: Contains the ResNet model implementation.
-"""
-
 from .base_model import VeryBasicModel, BasicModel, BasicClassifier
-from .resnet import ResNet
+from .mini_model import MiniCNNForTesting
 
-__all__ = ['VeryBasicModel', 'BasicModel', 'BasicClassifier', 'ResNet']
+__all__ = ['VeryBasicModel', 'BasicModel', 'BasicClassifier', 'MiniCNNForTesting']
